+++ docs/conf.py
@@ -5,29 +5,16 @@ import re
 ROOT = Path(__file__).parent.parent
 PACKAGE_SRC = ROOT / "jsonschema"
 
-# -- Project information -----------------------------------------------------
-
 project = "jsonschema"
 author = "Julian Berman"
 copyright = "2013, " + author
 
-# version: The short X.Y version
-# release: The full version, including alpha/beta/rc tags.
 release = importlib.metadata.version("jsonschema")
 version = release.partition("-")[0]
 
-
-# -- General configuration ---------------------------------------------------
-
-# If your documentation needs a minimal Sphinx version, state it here.
-#
-# needs_sphinx = "1.0"
-
+language = "en"
 default_role = "any"
 
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named "sphinx.ext.*") or your custom
-# ones.
 extensions = [
     "sphinx.ext.autodoc",
     "sphinx.ext.autosectionlabel",
@@ -36,9 +23,6 @@ extensions = [
     "sphinx.ext.intersphinx",
     "sphinx.ext.napoleon",
     "sphinx.ext.viewcode",
-
-    "autoapi.extension",
-    "sphinx_autodoc_typehints",
     "sphinx_copybutton",
     "sphinx_json_schema_spec",
     "sphinxcontrib.spelling",
@@ -47,99 +31,44 @@ extensions = [
 
 cache_path = "_cache"
 
-# Add any paths that contain templates here, relative to this directory.
-templates_path = ["_templates"]
-
-# The suffix(es) of source filenames.
-# You can specify multiple suffix as a list of string:
-#
-# source_suffix = [".rst", ".md"]
-source_suffix = ".rst"
-
-# The master toctree document.
-master_doc = "index"
-
-# There are two options for replacing |today|: either, you set today to some
-# non-false value, then it is used:
-# today = ""
-# Else, today_fmt is used as the format for a strftime call.
-# today_fmt = "%B %d, %Y"
-
-# List of patterns, relative to source directory, that match files and
-# directories to ignore when looking for source files.
-# This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ["_build", "_cache", "_static", "_templates"]
-
-# The name of the Pygments (syntax highlighting) style to use.
 pygments_style = "lovelace"
 pygments_dark_style = "one-dark"
 
-doctest_global_setup = """
-from jsonschema import *
-"""
+html_theme = "furo"
 
-intersphinx_mapping = {
-    "python": ("https://docs.python.org/3", None),
-    "ujs": ("https://json-schema.org/understanding-json-schema/", None),
+# See sphinx-doc/sphinx#10785
+_TYPE_ALIASES = {
+    "jsonschema._format._F",  # format checkers
 }
 
 
-# -- Options for HTML output -----------------------------------------------
-
-# The theme to use for HTML and HTML Help pages.  See the documentation for
-# a list of builtin themes.
-html_theme = "furo"
-
-# Theme options are theme-specific and customize the look and feel of a theme
-# further.  For a list of options available for each theme, see the
-# documentation.
-#
-# html_theme_options = {}
-
-# Add any paths that contain custom static files (such as style sheets) here,
-# relative to this directory. They are copied after the builtin static files,
-# so a file named "default.css" will overwrite the builtin "default.css".
-# html_static_path = ["_static"]
-
-
-# -- Options for HTMLHelp output ---------------------------------------------
-
-# Output file base name for HTML help builder.
-htmlhelp_basename = "jsonschemadoc"
+def _resolve_type_aliases(app, env, node, contnode):
+    if (
+        node["refdomain"] == "py"
+        and node["reftype"] == "class"
+        and node["reftarget"] in _TYPE_ALIASES
+    ):
+        return app.env.get_domain("py").resolve_xref(
+            env,
+            node["refdoc"],
+            app.builder,
+            "data",
+            node["reftarget"],
+            node,
+            contnode,
+        )
 
 
-# -- Options for LaTeX output ------------------------------------------------
-
-latex_documents = [
-    ("index", "jsonschema.tex", "jsonschema Documentation", author, "manual"),
-]
+def setup(app):
+    app.connect("missing-reference", _resolve_type_aliases)
 
 
-# -- Options for manual page output ------------------------------------------
+# = Builders =
 
-# One entry per manual page. List of tuples
-# (source start file, name, description, authors, manual section).
-man_pages = [("index", "jsonschema", "jsonschema Documentation", [author], 1)]
-
-
-# -- Options for Texinfo output ----------------------------------------------
-
-# Grouping the document tree into Texinfo files. List of tuples
-# (source start file, target name, title, author,
-#  dir menu entry, description, category)
-texinfo_documents = [
-    (
-        "index",
-        "jsonschema",
-        "jsonschema Documentation",
-        author,
-        "jsonschema",
-        "One line description of project.",
-        "Miscellaneous",
-    ),
-]
-
-# -- Options for the linkcheck builder --------------------------------------
+doctest_global_setup = """
+from jsonschema import *
+import jsonschema.validators
+"""
 
 
 def entire_domain(host):
@@ -152,19 +81,14 @@ linkcheck_ignore = [
     "https://github.com/python-jsonschema/jsonschema/workflows/CI/badge.svg",
 ]
 
-# -- Options for sphinxcontrib-autosectionlabel ---------------------------
+# = Extensions =
 
-autosectionlabel_prefix_document = True
-
-# -- Options for sphinxcontrib-spelling -----------------------------------
-
-spelling_word_list_filename = "spelling-wordlist.txt"
-
-# -- Options for autoapi ----------------------------------------------------
+# -- autoapi --
 
 suppress_warnings = [
     "autoapi.python_import_resolution",
     "autoapi.toc_reference",
+    "epub.duplicated_toc_entry",
 ]
 autoapi_root = "api"
 autoapi_ignore = [
@@ -183,3 +107,19 @@ autoapi_options = [
 
 autoapi_type = "python"
 autoapi_dirs = [PACKAGE_SRC]
+
+# -- autosectionlabel --
+
+autosectionlabel_prefix_document = True
+
+# -- intersphinx --
+
+intersphinx_mapping = {
+    "python": ("https://docs.python.org/3", None),
+    "ujs": ("https://json-schema.org/understanding-json-schema/", None),
+}
+
+# -- sphinxcontrib-spelling --
+
+spelling_word_list_filename = "spelling-wordlist.txt"
+spelling_show_suggestions = True