+++ jsonschema/cli.py
@@ -8,6 +8,7 @@ import argparse
 import json
 import sys
 import traceback
+import warnings
 
 try:
     from importlib import metadata
@@ -23,6 +24,16 @@ import attr
 
 from jsonschema.exceptions import SchemaError
 from jsonschema.validators import RefResolver, validator_for
+
+warnings.warn(
+    (
+        "The jsonschema CLI is deprecated and will be removed in a future "
+        "version. Please use check-jsonschema instead, which can be installed "
+        "from https://pypi.org/project/check-jsonschema/"
+    ),
+    DeprecationWarning,
+    stacklevel=2,
+)
 
 
 class _CannotLoadFile(Exception):
@@ -240,11 +251,12 @@ def run(arguments, stdout=sys.stdout, stderr=sys.stderr, stdin=sys.stdin):
     except _CannotLoadFile:
         return 1
 
-    if arguments["validator"] is None:
-        arguments["validator"] = validator_for(schema)
+    Validator = arguments["validator"]
+    if Validator is None:
+        Validator = validator_for(schema)
 
     try:
-        arguments["validator"].check_schema(schema)
+        Validator.check_schema(schema)
     except SchemaError as error:
         outputter.validation_error(
             instance_path=arguments["schema"],
@@ -270,7 +282,7 @@ def run(arguments, stdout=sys.stdout, stderr=sys.stderr, stdin=sys.stdin):
         referrer=schema,
     ) if arguments["base_uri"] is not None else None
 
-    validator = arguments["validator"](schema, resolver=resolver)
+    validator = Validator(schema, resolver=resolver)
     exit_code = 0
     for each in instances:
         try: