+++ jsonschema/tests/test_deprecations.py
@@ -1,4 +1,7 @@
 from unittest import TestCase
+import importlib
+import subprocess
+import sys
 
 from jsonschema import FormatChecker, validators
 
@@ -261,3 +264,30 @@ class TestDeprecations(TestCase):
 
         with self.assertRaises(ImportError):
             from jsonschema import draft1234_format_checker  # noqa
+
+    def test_import_cli(self):
+        """
+        As of v4.17.0, importing jsonschema.cli is deprecated.
+        """
+
+        with self.assertWarns(DeprecationWarning) as w:
+            import jsonschema.cli
+            importlib.reload(jsonschema.cli)
+
+        self.assertEqual(w.filename, importlib.__file__)
+        self.assertTrue(
+            str(w.warning).startswith(
+                "The jsonschema CLI is deprecated and will be removed ",
+            ),
+        )
+
+    def test_cli(self):
+        """
+        As of v4.17.0, the jsonschema CLI is deprecated.
+        """
+
+        process = subprocess.run(
+            [sys.executable, "-m", "jsonschema"],
+            capture_output=True,
+        )
+        self.assertIn(b"The jsonschema CLI is deprecated ", process.stderr)