+++ jsonschema/tests/test_jsonschema_test_suite.py
@@ -67,25 +67,6 @@ def complex_email_validation(test):
     )(test)
 
 
-is_narrow_build = sys.maxunicode == 2 ** 16 - 1
-if is_narrow_build:  # pragma: no cover
-    message = "Not running surrogate Unicode case, this Python is narrow."
-
-    def narrow_unicode_build(test):  # pragma: no cover
-        return skip(
-            message=message,
-            description=(
-                "one supplementary Unicode code point is not long enough"
-            ),
-        )(test) or skip(
-            message=message,
-            description="two supplementary Unicode code points is long enough",
-        )(test)
-else:
-    def narrow_unicode_build(test):  # pragma: no cover
-        return
-
-
 if sys.version_info < (3, 9):  # pragma: no cover
     message = "Rejecting leading zeros is 3.9+"
     allowed_leading_zeros = skip(
@@ -144,21 +125,19 @@ def leap_second(test):
 
 
 TestDraft3 = DRAFT3.to_unittest_testcase(
-    DRAFT3.tests(),
-    DRAFT3.format_tests(),
-    DRAFT3.optional_tests_of(name="bignum"),
-    DRAFT3.optional_tests_of(name="non-bmp-regex"),
-    DRAFT3.optional_tests_of(name="zeroTerminatedFloats"),
+    DRAFT3.cases(),
+    DRAFT3.format_cases(),
+    DRAFT3.optional_cases_of(name="bignum"),
+    DRAFT3.optional_cases_of(name="non-bmp-regex"),
+    DRAFT3.optional_cases_of(name="zeroTerminatedFloats"),
     Validator=jsonschema.Draft3Validator,
     format_checker=jsonschema.Draft3Validator.FORMAT_CHECKER,
     skip=lambda test: (
-        narrow_unicode_build(test)
-        or missing_format(jsonschema.Draft3Validator)(test)
+        missing_format(jsonschema.Draft3Validator)(test)
         or complex_email_validation(test)
         or skip(
             message=bug(),
             subject="ref",
-            valid=False,
             case_description=(
                 "$ref prevents a sibling id from changing the base uri"
             ),
@@ -168,17 +147,16 @@ TestDraft3 = DRAFT3.to_unittest_testcase(
 
 
 TestDraft4 = DRAFT4.to_unittest_testcase(
-    DRAFT4.tests(),
-    DRAFT4.format_tests(),
-    DRAFT4.optional_tests_of(name="bignum"),
-    DRAFT4.optional_tests_of(name="float-overflow"),
-    DRAFT4.optional_tests_of(name="non-bmp-regex"),
-    DRAFT4.optional_tests_of(name="zeroTerminatedFloats"),
+    DRAFT4.cases(),
+    DRAFT4.format_cases(),
+    DRAFT4.optional_cases_of(name="bignum"),
+    DRAFT4.optional_cases_of(name="float-overflow"),
+    DRAFT4.optional_cases_of(name="non-bmp-regex"),
+    DRAFT4.optional_cases_of(name="zeroTerminatedFloats"),
     Validator=jsonschema.Draft4Validator,
     format_checker=jsonschema.Draft4Validator.FORMAT_CHECKER,
     skip=lambda test: (
-        narrow_unicode_build(test)
-        or allowed_leading_zeros(test)
+        allowed_leading_zeros(test)
         or leap_second(test)
         or missing_format(jsonschema.Draft4Validator)(test)
         or complex_email_validation(test)
@@ -230,16 +208,15 @@ TestDraft4 = DRAFT4.to_unittest_testcase(
 
 
 TestDraft6 = DRAFT6.to_unittest_testcase(
-    DRAFT6.tests(),
-    DRAFT6.format_tests(),
-    DRAFT6.optional_tests_of(name="bignum"),
-    DRAFT6.optional_tests_of(name="float-overflow"),
-    DRAFT6.optional_tests_of(name="non-bmp-regex"),
+    DRAFT6.cases(),
+    DRAFT6.format_cases(),
+    DRAFT6.optional_cases_of(name="bignum"),
+    DRAFT6.optional_cases_of(name="float-overflow"),
+    DRAFT6.optional_cases_of(name="non-bmp-regex"),
     Validator=jsonschema.Draft6Validator,
     format_checker=jsonschema.Draft6Validator.FORMAT_CHECKER,
     skip=lambda test: (
-        narrow_unicode_build(test)
-        or allowed_leading_zeros(test)
+        allowed_leading_zeros(test)
         or leap_second(test)
         or missing_format(jsonschema.Draft6Validator)(test)
         or complex_email_validation(test)
@@ -253,17 +230,16 @@ TestDraft6 = DRAFT6.to_unittest_testcase(
 
 
 TestDraft7 = DRAFT7.to_unittest_testcase(
-    DRAFT7.tests(),
-    DRAFT7.format_tests(),
-    DRAFT7.optional_tests_of(name="bignum"),
-    DRAFT7.optional_tests_of(name="cross-draft"),
-    DRAFT7.optional_tests_of(name="float-overflow"),
-    DRAFT7.optional_tests_of(name="non-bmp-regex"),
+    DRAFT7.cases(),
+    DRAFT7.format_cases(),
+    DRAFT7.optional_cases_of(name="bignum"),
+    DRAFT7.optional_cases_of(name="cross-draft"),
+    DRAFT7.optional_cases_of(name="float-overflow"),
+    DRAFT7.optional_cases_of(name="non-bmp-regex"),
     Validator=jsonschema.Draft7Validator,
     format_checker=jsonschema.Draft7Validator.FORMAT_CHECKER,
     skip=lambda test: (
-        narrow_unicode_build(test)
-        or allowed_leading_zeros(test)
+        allowed_leading_zeros(test)
         or leap_second(test)
         or missing_format(jsonschema.Draft7Validator)(test)
         or complex_email_validation(test)
@@ -285,12 +261,12 @@ TestDraft7 = DRAFT7.to_unittest_testcase(
 
 
 TestDraft201909 = DRAFT201909.to_unittest_testcase(
-    DRAFT201909.tests(),
-    DRAFT201909.optional_tests_of(name="bignum"),
-    DRAFT201909.optional_tests_of(name="cross-draft"),
-    DRAFT201909.optional_tests_of(name="float-overflow"),
-    DRAFT201909.optional_tests_of(name="non-bmp-regex"),
-    DRAFT201909.optional_tests_of(name="refOfUnknownKeyword"),
+    DRAFT201909.cases(),
+    DRAFT201909.optional_cases_of(name="bignum"),
+    DRAFT201909.optional_cases_of(name="cross-draft"),
+    DRAFT201909.optional_cases_of(name="float-overflow"),
+    DRAFT201909.optional_cases_of(name="non-bmp-regex"),
+    DRAFT201909.optional_cases_of(name="refOfUnknownKeyword"),
     Validator=jsonschema.Draft201909Validator,
     skip=lambda test: (
         skip(
@@ -406,7 +382,8 @@ TestDraft201909 = DRAFT201909.to_unittest_testcase(
 
 
 TestDraft201909Format = DRAFT201909.to_unittest_testcase(
-    DRAFT201909.format_tests(),
+    DRAFT201909.format_cases(),
+    name="TestDraft201909Format",
     Validator=jsonschema.Draft201909Validator,
     format_checker=jsonschema.Draft201909Validator.FORMAT_CHECKER,
     skip=lambda test: (
@@ -420,16 +397,15 @@ TestDraft201909Format = DRAFT201909.to_unittest_testcase(
 
 
 TestDraft202012 = DRAFT202012.to_unittest_testcase(
-    DRAFT202012.tests(),
-    DRAFT202012.optional_tests_of(name="bignum"),
-    DRAFT202012.optional_tests_of(name="cross-draft"),
-    DRAFT202012.optional_tests_of(name="float-overflow"),
-    DRAFT202012.optional_tests_of(name="non-bmp-regex"),
-    DRAFT202012.optional_tests_of(name="refOfUnknownKeyword"),
+    DRAFT202012.cases(),
+    DRAFT202012.optional_cases_of(name="bignum"),
+    DRAFT202012.optional_cases_of(name="cross-draft"),
+    DRAFT202012.optional_cases_of(name="float-overflow"),
+    DRAFT202012.optional_cases_of(name="non-bmp-regex"),
+    DRAFT202012.optional_cases_of(name="refOfUnknownKeyword"),
     Validator=jsonschema.Draft202012Validator,
     skip=lambda test: (
-        narrow_unicode_build(test)
-        or skip(
+        skip(
             message="dynamicRef support isn't fully working yet.",
             subject="dynamicRef",
             description="The recursive part is not valid against the root",
@@ -531,7 +507,8 @@ TestDraft202012 = DRAFT202012.to_unittest_testcase(
 
 
 TestDraft202012Format = DRAFT202012.to_unittest_testcase(
-    DRAFT202012.format_tests(),
+    DRAFT202012.format_cases(),
+    name="TestDraft202012Format",
     Validator=jsonschema.Draft202012Validator,
     format_checker=jsonschema.Draft202012Validator.FORMAT_CHECKER,
     skip=lambda test: (